move remaining circular dependencies out of platform
