+++ packages/dd-trace/test/platform/node/index.spec.js
@@ -5,9 +5,6 @@ const os = require('os')
 const semver = require('semver')
 const { execSync } = require('child_process')
 
-const AgentExporter = require('../../../src/exporters/agent')
-const LogExporter = require('../../../src/exporters/log')
-
 wrapIt()
 
 describe('Platform', () => {
@@ -383,84 +380,6 @@ describe('Platform', () => {
         }, (err, res) => {
           expect(res).to.equal('OK')
         })
-      })
-    })
-
-    describe('getScope', () => {
-      let platform
-      let versionDescriptor
-      const ASYNC_LOCAL_STORAGE = { name: 'AsyncLocalStorage' }
-      const ASYNC_HOOKS = { name: 'async_hooks' }
-
-      beforeEach(() => {
-        versionDescriptor = Reflect.getOwnPropertyDescriptor(process.versions, 'node')
-      })
-
-      afterEach(() => {
-        Reflect.defineProperty(process.versions, 'node', versionDescriptor)
-      })
-
-      it('should default to AsyncLocalStorage on supported versions, and async_hooks on unsupported versions', () => {
-        function assertVersion (version, als) {
-          Reflect.defineProperty(process.versions, 'node', {
-            value: version,
-            configurable: true
-          })
-          platform = proxyquire('../src/platform/node/index', {
-            '../../scope/async_local_storage': ASYNC_LOCAL_STORAGE,
-            '../../scope/async_hooks': ASYNC_HOOKS
-          })
-          expect(platform.getScope()).to.equal(als ? ASYNC_LOCAL_STORAGE : ASYNC_HOOKS)
-        }
-        assertVersion('10.0.0', false)
-        assertVersion('12.0.0', false)
-        assertVersion('12.18.99', false)
-        assertVersion('12.19.0', true)
-        assertVersion('13.0.0', false)
-        assertVersion('14.0.0', false)
-        assertVersion('14.4.99', false)
-        assertVersion('14.5.0', true)
-        assertVersion('14.5.1', true)
-        assertVersion('14.6.0', true)
-        assertVersion('15.0.0', true)
-        assertVersion('16.0.0', true)
-        assertVersion('17.0.0', true)
-      })
-
-      it('should go with user choice when scope is defined in options', () => {
-        expect(platform.getScope('async_local_storage')).to.equal(ASYNC_LOCAL_STORAGE)
-        expect(platform.getScope('async_hooks')).to.equal(ASYNC_HOOKS)
-      })
-    })
-
-    describe('exporter', () => {
-      it('should create an AgentExporter by default', () => {
-        const Exporter = proxyquire('../src/platform/node/exporter', {
-          './env': () => undefined
-        })()
-
-        expect(Exporter).to.be.equal(AgentExporter)
-      })
-
-      it('should create an LogExporter when in Lambda environment', () => {
-        const Exporter = proxyquire('../src/platform/node/exporter', {
-          './env': (key) => {
-            if (key === 'AWS_LAMBDA_FUNCTION_NAME') {
-              return 'my-func'
-            }
-            return undefined
-          }
-        })()
-
-        expect(Exporter).to.be.equal(LogExporter)
-      })
-
-      it('should allow configuring the exporter', () => {
-        const Exporter = proxyquire('../src/platform/node/exporter', {
-          './env': () => undefined
-        })('log')
-
-        expect(Exporter).to.be.equal(LogExporter)
       })
     })
   })
